Moving widget interception logic to a utils fn.
